requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0
websocket-client>=1.6.4
cryptography>=41.0.0
httpx>=0.25.2
x10-python-trading-starknet
